requirements.txt
torch
torchvision
torchaudio
transformers
discord-py-interactions
python-dotenv